Reoptimized the Code and moved Static Data to an Apt Folder
