+++ frontend/package.json
@@ -17,6 +17,7 @@
     "react-multi-select-component": "^3.0.2",
     "react-router-dom": "^5.2.0",
     "react-scripts": "3.4.3",
+    "react-step-builder": "^1.1.11",
     "react-wordcloud": "^1.2.7"
   },
   "scripts": {
