+++ frontend/src/pages/data/CloudData.js
@@ -1,0 +1,68 @@
+const words = [
+  {
+    text: 'Python',
+    value: 75,
+  },
+  {
+    text: 'Git',
+    value: 55,
+  },
+  {
+    text: 'AWS',
+    value: 64,
+  },
+  {
+    text: 'API',
+    value: 33,
+  },
+  {
+    text: 'DynamoDB',
+    value: 43,
+  },
+  {
+    text: 'Elastic Beanstalk',
+    value: 26,
+  },
+  {
+    text: 'Github Actions',
+    value: 87,
+  },
+  {
+    text: 'AWS Lambda',
+    value: 50,
+  },
+  {
+    text: 'Neo4j',
+    value: 23,
+  },
+  {
+    text: 'MySQL',
+    value: 35,
+  },
+  {
+    text: 'Py2Neo',
+    value: 23,
+  },
+  {
+    text: 'SOAP',
+    value: 31,
+  },
+  {
+    text: 'Flask',
+    value: 12,
+  },
+  {
+    text: 'Django-REST',
+    value: 15,
+  },
+  {
+    text: 'GraphQL',
+    value: 30,
+  },
+  {
+    text: 'Selenium',
+    value: 18,
+  },
+]
+
+export default words;
