+++ frontend/src/pages/data/OverallData.js
@@ -1,0 +1,16 @@
+const OverallData = {
+  labels: ['Rohan Joshi', 'Mohan Kumar', 'Kiran CV', 'Akkul DN', 'Rajkumar', 'Smitha', 'Rohan Dudam'],
+  datasets: [
+    {
+      label: "Symmetry Score",
+      backgroundColor: 'LightBlue',
+      borderColor: 'black',
+      borderWidth: 1,
+      hoverBackgroundColor: 'blue',
+      hoverBorderColor: 'black',
+      data: [91, 100, 72, 91, 67, 66, 54]
+    }
+  ]
+};
+
+export default OverallData;
