+++ frontend/src/pages/data/SymmetryData.js
@@ -1,0 +1,16 @@
+const SymmetryData = {
+  labels: ['Shivahari', 'Arunkumar', 'Rahul', 'Raji', 'Avinash', 'Smitha', 'Rohit'],
+  datasets: [
+    {
+      label: "Symmetry Score",
+      backgroundColor: 'LightBlue',
+      borderColor: 'black',
+      borderWidth: 1,
+      hoverBackgroundColor: 'blue',
+      hoverBorderColor: 'black',
+      data: [234, 212, 110, 107, 67, 55, 16]
+    }
+  ]
+};
+
+export default SymmetryData;
