+++ frontend/src/pages/data/employeeList.js
@@ -1,0 +1,30 @@
+const options = [
+      { label: " Arunkumar Muralidharan", value: "mak@qxf3.com" },
+      { label: "Raji Gali", value: "raji@qxf3.com" },
+      { label: "Avinash Shetty", value: "avinash@qxf3.com" },
+      { label: "Shivahari Pitchaikkannu", value: "shivahari@qxf3.com" },
+      { label: "Annapoorani Gurusamy", value: "annapoorani@qxf3.com" },
+      { label: "Rohan Dudam", value: "rohan@qxf3.com" },
+      { label: "Smitha Rajesh", value: "smitha@qxf2.com" },
+      { label: "Indira Nellutla", value: "indira@qxf2.com" },
+      { label: "Rohan Joshi", value: "rohan.js@qxf2.com"},
+      { label: "Nilaya Indurkar", value: "nilaya@qxf2.com" },
+      { label: "Edward Alan", value: "edward@qxf2.com" },
+      { label: "Mohan Kumar", value: "mohan@qxf2.com" },
+      { label: "Kavitha Duraiswamy", value: "kavitha.d@qxf2.com" },
+      { label: "Rohini Gopal", value: "rohini.gopal@qxf2.com" },
+      { label: "Akkul DN", value: "akkul.dn@qxf2.com" },
+      { label: "Kiran CV", value: "kiran.cv@qxf2.com" },
+      { label: "Sravanti Tatiraju", value: "sravanti.tatiraju@qxf2.com" },
+      { label: "Rahul Bhave", value: "rahul.bhave@qxf2.com" },
+      { label: "Preedhi Vivek", value: "preedhi.vivek@qxf2.com" },
+      { label: "Drishya TM", value: "drishya.tm@qxf2.com" },
+      { label: "Harsh Bardhan", value: "harsh.bardhan@qxf2.com" },
+      { label: "Rohit Prasad", value: "rohit.prasad@qxf2.com" },
+      { label: "Basavaraj Hiremath", value: "basavaraj.hiremath@qxf2.com" },
+      { label: "Namitha Sathyananda", value: "namitha.sathyananda@qxf2.com" },
+      { label: "Vidhya Zierlein", value: "vidhya.zierlein@qxf2.com" },
+      { label: "Rakhi Chaudhuri", value: "rakhi.chaudhuri@qxf2.com" },
+    ];
+
+export default options;
